# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crypto Bot API",
        "description": "API para gerenciamento de portfólios de criptomoedas, estratégias de trading e alertas",
        "contact": {
            "name": "Suporte Crypto Bot",
            "email": "support@cryptobot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/strategies": {
            "get": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Listar estratégias",
                "description": "Listar todas as estratégias de um portfólio do usuário autenticado.",
                "operationId": "0726c656ef9f7473f4a36f3bc5bae6c2",
                "parameters": [
                    {
                        "name": "portfolio_id",
                        "in": "query",
                        "description": "ID do portfólio para filtrar estratégias",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo de estratégia (technical, fundamental, sentiment, hybrid)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "technical",
                                "fundamental",
                                "sentiment",
                                "hybrid"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por estratégias ativas",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estratégias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "portfolio_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "type": {
                                                "type": "string"
                                            },
                                            "parameters": {
                                                "type": "object"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            },
                                            "last_executed": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Criar estratégia",
                "description": "Criar uma nova estratégia em um portfólio do usuário autenticado.",
                "operationId": "2e150d271bbfa717051e211961563a3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "portfolio_id",
                                    "name",
                                    "type",
                                    "parameters"
                                ],
                                "properties": {
                                    "portfolio_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "technical",
                                            "fundamental",
                                            "sentiment",
                                            "hybrid"
                                        ]
                                    },
                                    "parameters": {
                                        "type": "object"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "schedule": {
                                        "type": "string"
                                    },
                                    "risk_level": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    },
                                    "max_positions": {
                                        "type": "integer"
                                    },
                                    "stop_loss": {
                                        "type": "number"
                                    },
                                    "take_profit": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estratégia criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "portfolio_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "parameters": {
                                            "type": "object"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}